Add an icon file
Same file as the 404, just shrunk down to 48 pixels and stretched a bit to
fill out the square canvadl (but it is still an svg). Someone with more design
skills can apply some spit and polish to it.

Note: like the .desktop file this ends up being installed in the egg when
using a normal setup.py run.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>