Conform to PEP8 style guide
It's good to have to code comply to a style guide; we go for PEP8
because it's the most used one. Checking for conformance should be done
using the `flake8` tool and the configuration given in this commit.

E402 and E501 are ignored because we don't care (that much) about line
length, and GObject Introspection requires us to import things as we
currently do.

piper/piper.py has not been fixed as it is going to be removed
eventually, when all its relevant bits have found a place in the new
code. It is therefore also ignored in the checking.